refactor(sql): mudando o tipo de nota_media para double

+++ sql_scripts/01_ddl_microdados.sql
@@ -2,6 +2,6 @@ DROP TABLE IF EXISTS microdados_enem_tratado;
 
 CREATE TABLE microdados_enem_tratado(
     id_estudante BIGINT PRIMARY KEY,
-    nota_media NUMERIC,
+    nota_media DOUBLE PRECISION,
     renda_categoria VARCHAR(50)
 );
